requirements.txt
wagtail==7.3.1
Django
anthropic
Pillow
python-dotenv
psycopg2-binary
whitenoise
